#13 ⚡️ fix : 간단한 수정
