+++ app/src/main/java/com/cokiri/coinkiri/presentation/post/CommunityDetail.kt
@@ -175,7 +175,7 @@ fun TitleSection() {
     }
 }
 
-@SuppressLint("SetJavaScriptEnabled")
+@SuppressLint("SetJavaScriptEnabled", "SuspiciousIndentation")
 @Composable
 fun ContentSection(content: String) {
     val context = LocalContext.current
